hosted: Moved all the raw SWD CMSIS-DAP routines into a new file
