+++ src/platforms/hosted/Makefile.inc
@@ -83,7 +83,7 @@ endif
 
 ifneq ($(HOSTED_BMP_ONLY), 1)
     CFLAGS += -DCMSIS_DAP
-    SRC += cmsis_dap.c dap.c dap_command.c
+    SRC += cmsis_dap.c dap.c dap_command.c dap_swd.c
     ifneq ($(shell pkg-config --exists $(HIDAPILIB); echo $$?), 0)
         $(error Please install $(HIDAPILIB) dependency or set HOSTED_BMP_ONLY to 1)
     endif
