+++ src/platforms/hosted/dap.c
@@ -473,71 +473,3 @@ bool dap_jtag_configure(void)
 		DEBUG_WARN("dap_jtag_configure failed with %02x\n", response);
 	return response == DAP_RESPONSE_OK;
 }
-
-void dap_swdptap_seq_out(const uint32_t tms_states, const size_t clock_cycles)
-{
-	/* Setup the sequence */
-	dap_swd_sequence_s sequence = {
-		clock_cycles,
-		DAP_SWD_OUT_SEQUENCE,
-	};
-	write_le4(sequence.data, 0, tms_states);
-	/* And perform it */
-	if (!perform_dap_swd_sequences(&sequence, 1U))
-		DEBUG_WARN("dap_swdptap_seq_out failed\n");
-}
-
-void dap_swdptap_seq_out_parity(const uint32_t tms_states, const size_t clock_cycles)
-{
-	/* Setup the sequence */
-	dap_swd_sequence_s sequence = {
-		clock_cycles + 1,
-		DAP_SWD_OUT_SEQUENCE,
-	};
-	write_le4(sequence.data, 0, tms_states);
-	sequence.data[4] = __builtin_parity(tms_states);
-	/* And perform it */
-	if (!perform_dap_swd_sequences(&sequence, 1U))
-		DEBUG_WARN("dap_swdptap_seq_out_parity failed\n");
-}
-
-uint32_t dap_swdptap_seq_in(const size_t clock_cycles)
-{
-	/* Setup the sequence */
-	dap_swd_sequence_s sequence = {
-		clock_cycles,
-		DAP_SWD_IN_SEQUENCE,
-	};
-	/* And perform it */
-	if (!perform_dap_swd_sequences(&sequence, 1U)) {
-		DEBUG_WARN("dap_swdptap_seq_in failed\n");
-		return 0U;
-	}
-
-	uint32_t result = 0;
-	for (size_t offset = 0; offset < clock_cycles; offset += 8U)
-		result |= sequence.data[offset >> 3U] << offset;
-	return result;
-}
-
-bool dap_swdptap_seq_in_parity(uint32_t *const result, const size_t clock_cycles)
-{
-	/* Setup the sequence */
-	dap_swd_sequence_s sequence = {
-		clock_cycles + 1U,
-		DAP_SWD_IN_SEQUENCE,
-	};
-	/* And perform it */
-	if (!perform_dap_swd_sequences(&sequence, 1U)) {
-		DEBUG_WARN("dap_swdptap_seq_in_parity failed\n");
-		return false;
-	}
-
-	uint32_t data = 0;
-	for (size_t offset = 0; offset < clock_cycles; offset += 8U)
-		data |= sequence.data[offset >> 3U] << offset;
-	*result = data;
-	uint8_t parity = __builtin_parity(data) & 1U;
-	parity ^= sequence.data[4] & 1U;
-	return !parity;
-}
